file-formula: don't try to share libmagic
A previous commit (aa601b0) modified the formula to attempt to share the
library and database from the libmagic formula. Unfortunately, that doesn't
seem to work in all environments. In particular, with file 5.28 on el_capitan
in /opt/homebrew, file cannot be run:

    dyld: Library not loaded: /opt/homebrew/Cellar/file-formula/5.28/lib/libmagic.1.dylib
      Referenced from: /opt/homebrew/Cellar/file-formula/5.28/bin/file
      Reason: image not found

It doesn't look like there is out-of-the-box support for this in libmagic/file.
Even if there was, it might be tricky getting the versions right. I think
ideally we'd have a single brew which was part-keg-only, but I don't see a way
to do that either. I'm open to suggestions.

Test plan:

`brew install --build-from-source file-formula && brew test file-formula`